Abstract Command types from Parser class
