+++ build.gradle
@@ -44,7 +44,7 @@ test {
 }
 
 application {
-    mainClassName = "duke.gui.Launcher"
+    mainClassName = "duke.Duke"
 }
 
 shadowJar {
